Refactor google maps direction and add geocode
