+++ src/main/kotlin/tech/jaya/ridely/integration/googlemaps/GoogleMapsFeignClient.kt
@@ -3,7 +3,8 @@ package tech.jaya.ridely.integration.googlemaps
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestParam
-import tech.jaya.ridely.integration.googlemaps.dto.GoogleMapsApiResponse
+import tech.jaya.ridely.integration.googlemaps.dto.GoogleMapsApiDirectionsResponse
+import tech.jaya.ridely.integration.googlemaps.dto.geocode.GoogleMapsApiGeocodeResponse
 
 @FeignClient(name = "googleMaps", url = "https://maps.googleapis.com")
 interface GoogleMapsFeignClient {
@@ -15,5 +16,12 @@ interface GoogleMapsFeignClient {
         @RequestParam destination: String,
         @RequestParam mode: String,
         @RequestParam key: String
-    ): GoogleMapsApiResponse
+    ): GoogleMapsApiDirectionsResponse
+
+    @GetMapping("/maps/api/geocode/json")
+    fun getGeoCode(
+        @RequestParam address: String,
+        @RequestParam mode: String,
+        @RequestParam key: String
+    ): GoogleMapsApiGeocodeResponse
 }
