--- src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/Distance.kt
@@ -1,3 +1,0 @@
-package tech.jaya.ridely.integration.googlemaps.dto
-
-data class Distance(val value: Int)
