--- src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/Duration.kt
@@ -1,3 +1,0 @@
-package tech.jaya.ridely.integration.googlemaps.dto
-
-data class Duration(val value: Int)
