--- src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/EndLocation.kt
@@ -1,3 +1,0 @@
-package tech.jaya.ridely.integration.googlemaps.dto
-
-data class EndLocation(val lat:Double, val lng: Double)
