+++ src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/GoogleMapsApiDirectionsResponse.kt
@@ -1,0 +1,7 @@
+package tech.jaya.ridely.integration.googlemaps.dto
+
+import tech.jaya.ridely.integration.googlemaps.dto.directions.Route
+
+data class GoogleMapsApiDirectionsResponse(val routes: List<Route>)
+
+
