--- src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/GoogleMapsApiResponse.kt
@@ -1,5 +1,0 @@
-package tech.jaya.ridely.integration.googlemaps.dto
-
-data class GoogleMapsApiResponse(val routes: List<Route>)
-
-
