--- src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/Route.kt
@@ -1,3 +1,0 @@
-package tech.jaya.ridely.integration.googlemaps.dto
-
-data class Route(val legs: List<Leg>)
