--- src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/StartLocation.kt
@@ -1,3 +1,0 @@
-package tech.jaya.ridely.integration.googlemaps.dto
-
-data class StartLocation(val lat:Double, val lng: Double)
