+++ src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/directions/Distance.kt
@@ -1,0 +1,3 @@
+package tech.jaya.ridely.integration.googlemaps.dto.directions
+
+data class Distance(val value: Int)
