+++ src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/directions/EndLocation.kt
@@ -1,0 +1,3 @@
+package tech.jaya.ridely.integration.googlemaps.dto.directions
+
+data class EndLocation(val lat:Double, val lng: Double)
