+++ src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/directions/Route.kt
@@ -1,0 +1,5 @@
+package tech.jaya.ridely.integration.googlemaps.dto.directions
+
+import tech.jaya.ridely.integration.googlemaps.dto.directions.Leg
+
+data class Route(val legs: List<Leg>)
