+++ src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/geocode/GoogleMapsApiGeocodeResponse.kt
@@ -1,0 +1,3 @@
+package tech.jaya.ridely.integration.googlemaps.dto.geocode
+
+data class GoogleMapsApiGeocodeResponse(val results: List<Result>)
