+++ src/main/kotlin/tech/jaya/ridely/integration/googlemaps/dto/geocode/Result.kt
@@ -1,0 +1,3 @@
+package tech.jaya.ridely.integration.googlemaps.dto.geocode
+
+data class Result(val geometry: Geometry)
